app1/requirements.txt
fastapi
uvicorn
redis
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-redis